+++ dcpp/ZUtils.cpp
@@ -164,7 +164,7 @@ void GZ::decompress(const string& source, const string& target) {
         if(read > 0) {
             f.write(&buf[0], read);
         }
-        if(read < BUF_SIZE) {
+        if(size_t(read) < BUF_SIZE) {
             break;
         }
     }
